[refactor] pull/push modal을 request modal로 통합
