+++ frontend/public/components/hypercloud/integration-config.tsx
@@ -11,7 +11,7 @@ import { IntegrationConfigModel } from '../../models/hypercloud';
 import { useTranslation } from 'react-i18next';
 import { TFunction } from 'i18next';
 import { ResourceLabel } from '../../models/hypercloud/resource-plural';
-import { integrationConfigPushModal, integrationConfigPullModal } from './modals';
+import { integrationConfigRequestModal } from './modals';
 import { asAccessReview } from '../utils/index';
 
 
@@ -23,8 +23,9 @@ export const menuActions: KebabAction[] = [...Kebab.getExtensionsActionsForKind(
   (kind, integrationConfig) => ({
     label: 'COMMON:MSG_MAIN_POPUP_20',
     callback: () =>
-      integrationConfigPushModal({
+      integrationConfigRequestModal({
         kind,
+        request: 'push',
         resource: integrationConfig,
         blocking: true,
       }),
@@ -33,8 +34,9 @@ export const menuActions: KebabAction[] = [...Kebab.getExtensionsActionsForKind(
   (kind, integrationConfig) => ({
     label: 'COMMON:MSG_MAIN_POPUP_23',
     callback: () =>
-      integrationConfigPullModal({
+      integrationConfigRequestModal({
         kind,
+        request: 'pull',
         resource: integrationConfig,
         blocking: true,
       }),
